requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlparse==0.5.1
tzdata==2024.2
djangorestframework-simplejwt
django-cors-headers
